Card Owner and modify endpoint

+++ webapp/src/main/java/ar/edu/itba/paw/cryptuki/form/UploadOfferForm.java
@@ -12,15 +12,15 @@ public class UploadOfferForm {
 
     @NotNull
     @DecimalMin("0.0000001")
-    private Float minAmount;
+    private Double minAmount;
 
     @NotNull
     @DecimalMin("0.00000001")
-    private Float maxAmount;
+    private Double maxAmount;
 
     @NotNull
     @DecimalMin("10.0")
-    private Float price;
+    private Double price;
 
     @Size(min=2, max=30)
     @NotNull
@@ -28,27 +28,27 @@ public class UploadOfferForm {
 
     @Size(min = 1)
     private String[] paymentMethods;
-    public Float getMinAmount() {
+    public Double getMinAmount() {
         return minAmount;
     }
 
-    public void setMinAmount(Float minAmount) {
+    public void setMinAmount(Double minAmount) {
         this.minAmount = minAmount;
     }
 
-    public void setMaxAmount(Float maxAmount) {
+    public void setMaxAmount(Double maxAmount) {
         this.maxAmount = maxAmount;
     }
 
-    public Float getMaxAmount() {
+    public Double getMaxAmount() {
         return maxAmount;
     }
 
-    public Float getPrice() {
+    public Double getPrice() {
         return price;
     }
 
-    public void setPrice(Float price) {
+    public void setPrice(Double price) {
         this.price = price;
     }
